+++ lib/shared/src/chat/prompts/utils.ts
@@ -94,3 +94,20 @@ export function getHumanTextForCommand(text: string, fileName?: string): string 
     }
     return promptText.replaceAll('{languageName}', getNormalizedLanguageName(getFileExtension(fileName)))
 }
+
+const leadingForwardSlashRegex = /^\/+/
+
+/**
+ * Removes leading forward slashes from slash command string.
+ */
+export function fromSlashCommand(slashCommand: string): string {
+    return slashCommand.replace(leadingForwardSlashRegex, '')
+}
+
+/**
+ * Returns command starting with a forward slash.
+ */
+export function toSlashCommand(command: string): string {
+    // ensure there is only one leading forward slash
+    return command.replace(leadingForwardSlashRegex, '').replace(/^/, '/')
+}