+++ vscode/src/chat/MessageProvider.ts
@@ -605,7 +605,7 @@ export abstract class MessageProvider extends MessageHandler implements vscode.D
                     await this.addCustomInteraction('Command failed. Please open a file and try again.', text)
                     return null
                 }
-                const promptText = this.editor.controllers.command?.find(text, true)
+                const promptText = this.editor.controllers.command?.find(text)
                 await this.editor.controllers.command?.get('command')
                 if (promptText) {
                     return { text: promptText, recipeId: 'custom-prompt' }
@@ -648,7 +648,7 @@ export abstract class MessageProvider extends MessageHandler implements vscode.D
     private async sendCodyCommands(): Promise<void> {
         const send = async (): Promise<void> => {
             await this.editor.controllers.command?.refresh()
-            const commands = (await this.editor.controllers.command?.getAllCommands()) || []
+            const commands = (await this.editor.controllers.command?.getAllCommands(true)) || []
             void this.handleCodyCommands(commands)
         }
         this.editor.controllers.command?.setMessenger(send)