+++ vscode/src/custom-prompts/PromptsProvider.ts
@@ -3,7 +3,7 @@ import { CodyPrompt, getDefaultCommandsMap } from '@sourcegraph/cody-shared/src/
 import { logDebug } from '../log'
 
 // Manage default commands created by the prompts in prompts.json
-const editorCommands = [{ name: 'Request a Code Edit', prompt: '/edit', slashCommand: '/edit' }]
+const editorCommands: CodyPrompt[] = [{ description: 'Edit code', prompt: '/edit', slashCommand: '/edit' }]
 export class PromptsProvider {
     // The default prompts
     private defaultPromptsMap = getDefaultCommandsMap(editorCommands)
@@ -19,13 +19,7 @@ export class PromptsProvider {
     /**
      * Find a prompt by its id
      */
-    public get(id: string, isSlashCommand = false): CodyPrompt | undefined {
-        if (id.startsWith('/') || isSlashCommand) {
-            const commands = [...this.allCommands]
-            const slashCommand = commands.find(command => command[1].slashCommand === id)
-            return slashCommand ? slashCommand[1] : this.allCommands.get(id)
-        }
-
+    public get(id: string): CodyPrompt | undefined {
         return this.allCommands.get(id)
     }
 
@@ -44,7 +38,7 @@ export class PromptsProvider {
      */
     public groupCommands(customCommands = new Map<string, CodyPrompt>()): void {
         const combinedMap = new Map([...this.defaultPromptsMap])
-        combinedMap.set('separator', { prompt: 'separator' })
+        combinedMap.set('separator', { prompt: 'separator', slashCommand: '' })
         this.allCommands = new Map([...combinedMap, ...customCommands])
     }